[active-active] Add support to send/handle mux probe request (#147)
Approach
What is the motivation for this PR?
For issue #143, if there is a link drop for upstream traffic, the ToR will try to toggle both itself and peer to standby. But the peer ToR might still be able to receive heartbeats and should be in active state.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Enable linkmgrd to support sending mux probe command to its peer when it tries to toggle peer to standby.

How did you verify/test it?
Add unittests and verify on real dualtor pairs.