[active-active] Add mux server state correction mechanism
There is no uniform mux server state correction for the state machine in
case if the mux toggle fails at first(mux server state not expected).
Let's add an uniform mux server state correction that is compatible with
previous codes.

The basic idea is that, when the mux probe is able to return meaningful
mux states(`active` or `standby`), let's add check to see if last toggle
succeeded or not. So let's define our mux toggle decision process as the
following:

```
If mux config is active or standby:
    if last toggle success:
        NOOP
    else:
        toggle based on the mux config
else:
    if mux status not matches heartbeats:
        toggle based on heartbeats
    else:
        if last toggle success:
            NOOP
        else:
            toggle based on the heartbeats
```

As we only consider mux probe `active` or `standby`, so only four
transition functions needs to be improved:

LinkProberActiveMuxActiveLinkUpTransitionFunction
LinkProberActiveMuxStandbyLinkUpTransitionFunction
LinkProberUnknownMuxActiveLinkUpTransitionFunction
LinkProberUnknownMuxStandbyLinkUpTransitionFunction

NOTE: mux probe `unknown` is not covered as the primary goal is to
guarantee a toggle success. Mux server down after toggle is not a
sufficient reason for us to update the mux server status and break the
healthy state.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>