[codeql] Fix codeql issues  (#205)
Description of PR
Summary:
Fixes # (issue)

Type of change
 Bug fix

Approach
What is the motivation for this PR?
use pull_request as trigger to include PR changes in analysis.

NOTE: pull_request_target doesn't include PR changes in the analysis, it is pointless to do so in PR testing.
ensure link directory(-L) is added before relative library(-l) in
linking.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
As the motivation