Fix mux port `uninitialized`
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] New feature
- [ ] Doc/Design
- [ ] Unit test

### Approach
#### What is the motivation for this PR?
Fix #307

The mux port health is `uninitialized` in nightly.
And from the syslog, the state machines are not activated as the link probers are trying to generate `GUID`. The `GUID` generation logic is stuck somehow in the multi-threading environment.
From the backtrace on each threads, some threads is busy looping the generated `GUID` set while some threads are trying to insert new `GUID` into the set; this is a race condition as the C++ multiset is not thread-safe.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>

##### Work item tracking
- Microsoft ADO **(number only)**: 34335369

#### How did you do it?
Let's wrap the generated `GUID` set update/read code section with a lock.

#### How did you verify/test it?
1. UT.
2. use the script in the issue, run over 12h, no repro.

#### Any platform specific information?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->